+++ library/UnitTests/Collections/TestHexCollection.cpp
@@ -4310,14 +4310,9 @@ BOOST_AUTO_TEST_CASE(TestHexHelmholtz_NoCollection_UniformP)
     Array<OneD, NekDouble> coeffsRef(nelmts * nm);
     Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
 
-    for (int i = 0; i < nm; ++i)
+    for (int i = 0; i < coeffsIn.size(); ++i)
     {
-        coeffsIn[i] = 1.0;
-    }
-
-    for (int i = 1; i < nelmts; ++i)
-    {
-        Vmath::Vcopy(nm, coeffsIn, 1, tmp = coeffsIn + i * nm, 1);
+        coeffsIn[i] = i + 1.0;
     }
 
     StdRegions::StdMatrixKey mkey(StdRegions::eHelmholtz, Exp->DetShapeType(),
@@ -4408,14 +4403,9 @@ BOOST_AUTO_TEST_CASE(TestHexHelmholtz_IterPerExp_UniformP)
     Array<OneD, NekDouble> coeffsRef(nelmts * nm);
     Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
 
-    for (int i = 0; i < nm; ++i)
+    for (int i = 0; i < coeffsIn.size(); ++i)
     {
-        coeffsIn[i] = 1.0;
-    }
-
-    for (int i = 1; i < nelmts; ++i)
-    {
-        Vmath::Vcopy(nm, coeffsIn, 1, tmp = coeffsIn + i * nm, 1);
+        coeffsIn[i] = i + 1.0;
     }
 
     StdRegions::StdMatrixKey mkey(StdRegions::eHelmholtz, Exp->DetShapeType(),
@@ -4512,14 +4502,9 @@ BOOST_AUTO_TEST_CASE(TestHexHelmholtz_IterPerExp_UniformP_ConstVarDiff)
     Array<OneD, NekDouble> coeffsRef(nelmts * nm);
     Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
 
-    for (int i = 0; i < nm; ++i)
+    for (int i = 0; i < coeffsIn.size(); ++i)
     {
-        coeffsIn[i] = 1.0;
-    }
-
-    for (int i = 1; i < nelmts; ++i)
-    {
-        Vmath::Vcopy(nm, coeffsIn, 1, tmp = coeffsIn + i * nm, 1);
+        coeffsIn[i] = i + 1.0;
     }
 
     StdRegions::StdMatrixKey mkey(StdRegions::eHelmholtz, Exp->DetShapeType(),
@@ -4601,7 +4586,7 @@ BOOST_AUTO_TEST_CASE(TestHexHelmholtz_MatrixFree_UniformP)
     Collections::OperatorImpMap impTypes = colOpt.GetOperatorImpMap(stdExp);
     Collections::Collection c(CollExp, impTypes);
     StdRegions::ConstFactorMap factors;
-    factors[StdRegions::eFactorLambda] = 0.0;
+    factors[StdRegions::eFactorLambda] = 1.5;
 
     c.Initialise(Collections::eHelmholtz, factors);
 
@@ -4610,14 +4595,9 @@ BOOST_AUTO_TEST_CASE(TestHexHelmholtz_MatrixFree_UniformP)
     Array<OneD, NekDouble> coeffsRef(nelmts * nm);
     Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
 
-    for (int i = 0; i < nm; ++i)
+    for (int i = 0; i < coeffsIn.size(); ++i)
     {
-        coeffsIn[i] = 1.0;
-    }
-
-    for (int i = 1; i < nelmts; ++i)
-    {
-        Vmath::Vcopy(nm, coeffsIn, 1, tmp = coeffsIn + i * nm, 1);
+        coeffsIn[i] = i + 1.0;
     }
 
     StdRegions::StdMatrixKey mkey(StdRegions::eHelmholtz, Exp->DetShapeType(),
@@ -4699,7 +4679,7 @@ BOOST_AUTO_TEST_CASE(TestHexHelmholtz_MatrixFree_UniformP_Deformed_OverInt)
     Collections::OperatorImpMap impTypes = colOpt.GetOperatorImpMap(stdExp);
     Collections::Collection c(CollExp, impTypes);
     StdRegions::ConstFactorMap factors;
-    factors[StdRegions::eFactorLambda] = 0.0;
+    factors[StdRegions::eFactorLambda] = 1.5;
 
     c.Initialise(Collections::eHelmholtz, factors);
 
@@ -4708,14 +4688,9 @@ BOOST_AUTO_TEST_CASE(TestHexHelmholtz_MatrixFree_UniformP_Deformed_OverInt)
     Array<OneD, NekDouble> coeffsRef(nelmts * nm);
     Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
 
-    for (int i = 0; i < nm; ++i)
+    for (int i = 0; i < coeffsIn.size(); ++i)
     {
-        coeffsIn[i] = 1.0;
-    }
-
-    for (int i = 1; i < nelmts; ++i)
-    {
-        Vmath::Vcopy(nm, coeffsIn, 1, tmp = coeffsIn + i * nm, 1);
+        coeffsIn[i] = i + 1.0;
     }
 
     StdRegions::StdMatrixKey mkey(StdRegions::eHelmholtz, Exp->DetShapeType(),
@@ -4797,7 +4772,7 @@ BOOST_AUTO_TEST_CASE(TestHexHelmholtz_MatrixFree_UniformP_ConstVarDiff)
     Collections::OperatorImpMap impTypes = colOpt.GetOperatorImpMap(stdExp);
     Collections::Collection c(CollExp, impTypes);
     StdRegions::ConstFactorMap factors;
-    factors[StdRegions::eFactorLambda]   = 1.0;
+    factors[StdRegions::eFactorLambda]   = 1.5;
     factors[StdRegions::eFactorCoeffD00] = 1.25;
     factors[StdRegions::eFactorCoeffD01] = 0.25;
     factors[StdRegions::eFactorCoeffD11] = 1.25;
@@ -4812,14 +4787,9 @@ BOOST_AUTO_TEST_CASE(TestHexHelmholtz_MatrixFree_UniformP_ConstVarDiff)
     Array<OneD, NekDouble> coeffsRef(nelmts * nm);
     Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
 
-    for (int i = 0; i < nm; ++i)
+    for (int i = 0; i < coeffsIn.size(); ++i)
     {
-        coeffsIn[i] = 1.0;
-    }
-
-    for (int i = 1; i < nelmts; ++i)
-    {
-        Vmath::Vcopy(nm, coeffsIn, 1, tmp = coeffsIn + i * nm, 1);
+        coeffsIn[i] = i + 1.0;
     }
 
     StdRegions::StdMatrixKey mkey(StdRegions::eHelmholtz, Exp->DetShapeType(),
@@ -5023,5 +4993,446 @@ BOOST_AUTO_TEST_CASE(TestHexPhysInterp1D_MatrixFree_UniformP)
         BOOST_CHECK_CLOSE(phys1[i], exact, epsilon);
     }
 }
+
+BOOST_AUTO_TEST_CASE(
+    TestHexLinearAdvectionDiffusionReaction_NoCollection_UniformP)
+{
+    SpatialDomains::PointGeomUniquePtr v0(
+        new SpatialDomains::PointGeom(3u, 0u, -1.0, -1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v1(
+        new SpatialDomains::PointGeom(3u, 1u, 1.0, -1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v2(
+        new SpatialDomains::PointGeom(3u, 2u, 1.0, 1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v3(
+        new SpatialDomains::PointGeom(3u, 3u, -1.0, 1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v4(
+        new SpatialDomains::PointGeom(3u, 4u, -1.0, -1.0, 1.0));
+    SpatialDomains::PointGeomUniquePtr v5(
+        new SpatialDomains::PointGeom(3u, 5u, 1.0, -1.0, 1.0));
+    SpatialDomains::PointGeomUniquePtr v6(
+        new SpatialDomains::PointGeom(3u, 6u, 2.0, 3.0, 4.0));
+    SpatialDomains::PointGeomUniquePtr v7(
+        new SpatialDomains::PointGeom(3u, 7u, -1.0, 1.0, 1.0));
+
+    std::array<SpatialDomains::PointGeom *, 8> v = {
+        v0.get(), v1.get(), v2.get(), v3.get(),
+        v4.get(), v5.get(), v6.get(), v7.get()};
+    std::array<SpatialDomains::SegGeomUniquePtr, 12> segVec;
+    std::array<SpatialDomains::QuadGeomUniquePtr, 6> faceVec;
+    SpatialDomains::HexGeomUniquePtr hexGeom = CreateHex(v, segVec, faceVec);
+
+    Nektar::LibUtilities::PointsType quadPointsTypeDir1 =
+        Nektar::LibUtilities::eGaussLobattoLegendre;
+    Nektar::LibUtilities::BasisType basisTypeDir1 =
+        Nektar::LibUtilities::eModified_A;
+    unsigned int numQuadPoints = 5;
+    unsigned int numModes      = 4;
+
+    const Nektar::LibUtilities::PointsKey quadPointsKeyDir1(numQuadPoints,
+                                                            quadPointsTypeDir1);
+    const Nektar::LibUtilities::BasisKey basisKeyDir1(basisTypeDir1, numModes,
+                                                      quadPointsKeyDir1);
+
+    Nektar::LocalRegions::HexExpSharedPtr Exp =
+        MemoryManager<Nektar::LocalRegions::HexExp>::AllocateSharedPtr(
+            basisKeyDir1, basisKeyDir1, basisKeyDir1, hexGeom.get());
+
+    Nektar::StdRegions::StdHexExpSharedPtr stdExp =
+        MemoryManager<Nektar::StdRegions::StdHexExp>::AllocateSharedPtr(
+            basisKeyDir1, basisKeyDir1, basisKeyDir1);
+
+    int nelmts = 10;
+
+    std::vector<StdRegions::StdExpansionSharedPtr> CollExp;
+    for (int i = 0; i < nelmts; ++i)
+    {
+        CollExp.push_back(Exp);
+    }
+
+    LibUtilities::SessionReaderSharedPtr dummySession;
+    Collections::CollectionOptimisation colOpt(dummySession, 2,
+                                               Collections::eNoCollection);
+    Collections::OperatorImpMap impTypes = colOpt.GetOperatorImpMap(stdExp);
+    Collections::Collection c(CollExp, impTypes);
+    StdRegions::ConstFactorMap factors;
+    factors[StdRegions::eFactorLambda] = 1.5;
+
+    c.Initialise(Collections::eLinearAdvectionDiffusionReaction, factors);
+
+    // Add advection velocities via varcoeffs
+    int npoints = Exp->GetTotPoints() * nelmts;
+    StdRegions::VarCoeffMap varcoeffs;
+    StdRegions::VarCoeffType varcoefftypes[] = {StdRegions::eVarCoeffVelX,
+                                                StdRegions::eVarCoeffVelY,
+                                                StdRegions::eVarCoeffVelZ};
+    for (int i = 0; i < Exp->GetShapeDimension(); i++)
+    {
+        varcoeffs[varcoefftypes[i]] = Array<OneD, NekDouble>(npoints, 1.0);
+    }
+    c.UpdateVarcoeffs(Collections::eLinearAdvectionDiffusionReaction,
+                      varcoeffs);
+
+    const int nm = Exp->GetNcoeffs();
+    Array<OneD, NekDouble> coeffsIn(nelmts * nm);
+    Array<OneD, NekDouble> coeffsRef(nelmts * nm);
+    Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
+
+    for (int i = 0; i < coeffsIn.size(); ++i)
+    {
+        coeffsIn[i] = i + 1.0;
+    }
+
+    StdRegions::StdMatrixKey mkey(StdRegions::eLinearAdvectionDiffusionReaction,
+                                  Exp->DetShapeType(), *Exp, factors,
+                                  varcoeffs);
+
+    for (int i = 0; i < nelmts; ++i)
+    {
+        // Standard routines
+        Exp->GeneralMatrixOp(coeffsIn + i * nm, tmp = coeffsRef + i * nm, mkey);
+    }
+
+    c.ApplyOperator(Collections::eLinearAdvectionDiffusionReaction, coeffsIn,
+                    coeffs);
+
+    double epsilon = 1.0e-8;
+    for (int i = 0; i < coeffsRef.size(); ++i)
+    {
+        coeffsRef[i] = (std::abs(coeffsRef[i]) < 1e-14) ? 0.0 : coeffsRef[i];
+        coeffs[i]    = (std::abs(coeffs[i]) < 1e-14) ? 0.0 : coeffs[i];
+        BOOST_CHECK_CLOSE(coeffsRef[i], coeffs[i], epsilon);
+    }
+}
+
+BOOST_AUTO_TEST_CASE(
+    TestHexLinearAdvectionDiffusionReaction_IterPerExp_UniformP)
+{
+    SpatialDomains::PointGeomUniquePtr v0(
+        new SpatialDomains::PointGeom(3u, 0u, -1.0, -1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v1(
+        new SpatialDomains::PointGeom(3u, 1u, 1.0, -1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v2(
+        new SpatialDomains::PointGeom(3u, 2u, 1.0, 1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v3(
+        new SpatialDomains::PointGeom(3u, 3u, -1.0, 1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v4(
+        new SpatialDomains::PointGeom(3u, 4u, -1.0, -1.0, 1.0));
+    SpatialDomains::PointGeomUniquePtr v5(
+        new SpatialDomains::PointGeom(3u, 5u, 1.0, -1.0, 1.0));
+    SpatialDomains::PointGeomUniquePtr v6(
+        new SpatialDomains::PointGeom(3u, 6u, 2.0, 3.0, 4.0));
+    SpatialDomains::PointGeomUniquePtr v7(
+        new SpatialDomains::PointGeom(3u, 7u, -1.0, 1.0, 1.0));
+
+    std::array<SpatialDomains::PointGeom *, 8> v = {
+        v0.get(), v1.get(), v2.get(), v3.get(),
+        v4.get(), v5.get(), v6.get(), v7.get()};
+    std::array<SpatialDomains::SegGeomUniquePtr, 12> segVec;
+    std::array<SpatialDomains::QuadGeomUniquePtr, 6> faceVec;
+    SpatialDomains::HexGeomUniquePtr hexGeom = CreateHex(v, segVec, faceVec);
+
+    Nektar::LibUtilities::PointsType quadPointsTypeDir1 =
+        Nektar::LibUtilities::eGaussLobattoLegendre;
+    Nektar::LibUtilities::BasisType basisTypeDir1 =
+        Nektar::LibUtilities::eModified_A;
+    unsigned int numQuadPoints = 5;
+    unsigned int numModes      = 4;
+
+    const Nektar::LibUtilities::PointsKey quadPointsKeyDir1(numQuadPoints,
+                                                            quadPointsTypeDir1);
+    const Nektar::LibUtilities::BasisKey basisKeyDir1(basisTypeDir1, numModes,
+                                                      quadPointsKeyDir1);
+
+    Nektar::LocalRegions::HexExpSharedPtr Exp =
+        MemoryManager<Nektar::LocalRegions::HexExp>::AllocateSharedPtr(
+            basisKeyDir1, basisKeyDir1, basisKeyDir1, hexGeom.get());
+
+    Nektar::StdRegions::StdHexExpSharedPtr stdExp =
+        MemoryManager<Nektar::StdRegions::StdHexExp>::AllocateSharedPtr(
+            basisKeyDir1, basisKeyDir1, basisKeyDir1);
+
+    int nelmts = 10;
+
+    std::vector<StdRegions::StdExpansionSharedPtr> CollExp;
+    for (int i = 0; i < nelmts; ++i)
+    {
+        CollExp.push_back(Exp);
+    }
+
+    LibUtilities::SessionReaderSharedPtr dummySession;
+    Collections::CollectionOptimisation colOpt(dummySession, 2,
+                                               Collections::eIterPerExp);
+    Collections::OperatorImpMap impTypes = colOpt.GetOperatorImpMap(stdExp);
+    Collections::Collection c(CollExp, impTypes);
+    StdRegions::ConstFactorMap factors;
+    factors[StdRegions::eFactorLambda] = 1.5;
+
+    c.Initialise(Collections::eLinearAdvectionDiffusionReaction, factors);
+
+    // Add advection velocities via varcoeffs
+    int npoints = Exp->GetTotPoints() * nelmts;
+    StdRegions::VarCoeffMap varcoeffs;
+    StdRegions::VarCoeffType varcoefftypes[] = {StdRegions::eVarCoeffVelX,
+                                                StdRegions::eVarCoeffVelY,
+                                                StdRegions::eVarCoeffVelZ};
+    for (int i = 0; i < Exp->GetShapeDimension(); i++)
+    {
+        varcoeffs[varcoefftypes[i]] = Array<OneD, NekDouble>(npoints, 1.0);
+    }
+    c.UpdateVarcoeffs(Collections::eLinearAdvectionDiffusionReaction,
+                      varcoeffs);
+
+    const int nm = Exp->GetNcoeffs();
+    Array<OneD, NekDouble> coeffsIn(nelmts * nm);
+    Array<OneD, NekDouble> coeffsRef(nelmts * nm);
+    Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
+
+    for (int i = 0; i < coeffsIn.size(); ++i)
+    {
+        coeffsIn[i] = i + 1.0;
+    }
+
+    StdRegions::StdMatrixKey mkey(StdRegions::eLinearAdvectionDiffusionReaction,
+                                  Exp->DetShapeType(), *Exp, factors,
+                                  varcoeffs);
+
+    for (int i = 0; i < nelmts; ++i)
+    {
+        // Standard routines
+        Exp->GeneralMatrixOp(coeffsIn + i * nm, tmp = coeffsRef + i * nm, mkey);
+    }
+
+    c.ApplyOperator(Collections::eLinearAdvectionDiffusionReaction, coeffsIn,
+                    coeffs);
+
+    double epsilon = 1.0e-8;
+    for (int i = 0; i < coeffsRef.size(); ++i)
+    {
+        coeffsRef[i] = (std::abs(coeffsRef[i]) < 1e-14) ? 0.0 : coeffsRef[i];
+        coeffs[i]    = (std::abs(coeffs[i]) < 1e-14) ? 0.0 : coeffs[i];
+        BOOST_CHECK_CLOSE(coeffsRef[i], coeffs[i], epsilon);
+    }
+}
+
+BOOST_AUTO_TEST_CASE(
+    TestHexLinearAdvectionDiffusionReaction_MatrixFree_UniformP)
+{
+    SpatialDomains::PointGeomUniquePtr v0(
+        new SpatialDomains::PointGeom(3u, 0u, -1.0, -1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v1(
+        new SpatialDomains::PointGeom(3u, 1u, 1.0, -1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v2(
+        new SpatialDomains::PointGeom(3u, 2u, 1.0, 1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v3(
+        new SpatialDomains::PointGeom(3u, 3u, -1.0, 1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v4(
+        new SpatialDomains::PointGeom(3u, 4u, -1.0, -1.0, 1.0));
+    SpatialDomains::PointGeomUniquePtr v5(
+        new SpatialDomains::PointGeom(3u, 5u, 1.0, -1.0, 1.0));
+    SpatialDomains::PointGeomUniquePtr v6(
+        new SpatialDomains::PointGeom(3u, 6u, 2.0, 3.0, 4.0));
+    SpatialDomains::PointGeomUniquePtr v7(
+        new SpatialDomains::PointGeom(3u, 7u, -1.0, 1.0, 1.0));
+
+    std::array<SpatialDomains::PointGeom *, 8> v = {
+        v0.get(), v1.get(), v2.get(), v3.get(),
+        v4.get(), v5.get(), v6.get(), v7.get()};
+    std::array<SpatialDomains::SegGeomUniquePtr, 12> segVec;
+    std::array<SpatialDomains::QuadGeomUniquePtr, 6> faceVec;
+    SpatialDomains::HexGeomUniquePtr hexGeom = CreateHex(v, segVec, faceVec);
+
+    Nektar::LibUtilities::PointsType quadPointsTypeDir1 =
+        Nektar::LibUtilities::eGaussLobattoLegendre;
+    Nektar::LibUtilities::BasisType basisTypeDir1 =
+        Nektar::LibUtilities::eModified_A;
+    unsigned int numQuadPoints = 5;
+    unsigned int numModes      = 4;
+
+    const Nektar::LibUtilities::PointsKey quadPointsKeyDir1(numQuadPoints,
+                                                            quadPointsTypeDir1);
+    const Nektar::LibUtilities::BasisKey basisKeyDir1(basisTypeDir1, numModes,
+                                                      quadPointsKeyDir1);
+
+    Nektar::LocalRegions::HexExpSharedPtr Exp =
+        MemoryManager<Nektar::LocalRegions::HexExp>::AllocateSharedPtr(
+            basisKeyDir1, basisKeyDir1, basisKeyDir1, hexGeom.get());
+
+    Nektar::StdRegions::StdHexExpSharedPtr stdExp =
+        MemoryManager<Nektar::StdRegions::StdHexExp>::AllocateSharedPtr(
+            basisKeyDir1, basisKeyDir1, basisKeyDir1);
+
+    int nelmts = 10;
+
+    std::vector<StdRegions::StdExpansionSharedPtr> CollExp;
+    for (int i = 0; i < nelmts; ++i)
+    {
+        CollExp.push_back(Exp);
+    }
+
+    LibUtilities::SessionReaderSharedPtr dummySession;
+    Collections::CollectionOptimisation colOpt(dummySession, 2,
+                                               Collections::eMatrixFree);
+    Collections::OperatorImpMap impTypes = colOpt.GetOperatorImpMap(stdExp);
+    Collections::Collection c(CollExp, impTypes);
+    StdRegions::ConstFactorMap factors;
+    factors[StdRegions::eFactorLambda] = 1.5;
+
+    c.Initialise(Collections::eLinearAdvectionDiffusionReaction, factors);
+
+    // Add advection velocities via varcoeffs
+    int npoints = Exp->GetTotPoints() * nelmts;
+    StdRegions::VarCoeffMap varcoeffs;
+    StdRegions::VarCoeffType varcoefftypes[] = {StdRegions::eVarCoeffVelX,
+                                                StdRegions::eVarCoeffVelY,
+                                                StdRegions::eVarCoeffVelZ};
+    for (int i = 0; i < Exp->GetShapeDimension(); i++)
+    {
+        varcoeffs[varcoefftypes[i]] = Array<OneD, NekDouble>(npoints, 1.0);
+    }
+    c.UpdateVarcoeffs(Collections::eLinearAdvectionDiffusionReaction,
+                      varcoeffs);
+
+    const int nm = Exp->GetNcoeffs();
+    Array<OneD, NekDouble> coeffsIn(nelmts * nm);
+    Array<OneD, NekDouble> coeffsRef(nelmts * nm);
+    Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
+
+    for (int i = 0; i < coeffsIn.size(); ++i)
+    {
+        coeffsIn[i] = i + 1.0;
+    }
+
+    StdRegions::StdMatrixKey mkey(StdRegions::eLinearAdvectionDiffusionReaction,
+                                  Exp->DetShapeType(), *Exp, factors,
+                                  varcoeffs);
+
+    for (int i = 0; i < nelmts; ++i)
+    {
+        // Standard routines
+        Exp->GeneralMatrixOp(coeffsIn + i * nm, tmp = coeffsRef + i * nm, mkey);
+    }
+
+    c.ApplyOperator(Collections::eLinearAdvectionDiffusionReaction, coeffsIn,
+                    coeffs);
+
+    double epsilon = 1.0e-8;
+    for (int i = 0; i < coeffsRef.size(); ++i)
+    {
+        coeffsRef[i] = (std::abs(coeffsRef[i]) < 1e-14) ? 0.0 : coeffsRef[i];
+        coeffs[i]    = (std::abs(coeffs[i]) < 1e-14) ? 0.0 : coeffs[i];
+        BOOST_CHECK_CLOSE(coeffsRef[i], coeffs[i], epsilon);
+    }
+}
+
+BOOST_AUTO_TEST_CASE(
+    TestHexLinearAdvectionDiffusionReaction_MatrixFree_UniformP_Deformed_OverInt)
+{
+    SpatialDomains::PointGeomUniquePtr v0(
+        new SpatialDomains::PointGeom(3u, 0u, -1.0, -1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v1(
+        new SpatialDomains::PointGeom(3u, 1u, 1.0, -1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v2(
+        new SpatialDomains::PointGeom(3u, 2u, 1.0, 1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v3(
+        new SpatialDomains::PointGeom(3u, 3u, -1.0, 1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v4(
+        new SpatialDomains::PointGeom(3u, 4u, -1.0, -1.0, 1.0));
+    SpatialDomains::PointGeomUniquePtr v5(
+        new SpatialDomains::PointGeom(3u, 5u, 1.0, -1.0, 1.0));
+    SpatialDomains::PointGeomUniquePtr v6(
+        new SpatialDomains::PointGeom(3u, 6u, 2.0, 3.0, 4.0));
+    SpatialDomains::PointGeomUniquePtr v7(
+        new SpatialDomains::PointGeom(3u, 7u, -1.0, 1.0, 1.0));
+
+    std::array<SpatialDomains::PointGeom *, 8> v = {
+        v0.get(), v1.get(), v2.get(), v3.get(),
+        v4.get(), v5.get(), v6.get(), v7.get()};
+    std::array<SpatialDomains::SegGeomUniquePtr, 12> segVec;
+    std::array<SpatialDomains::QuadGeomUniquePtr, 6> faceVec;
+    SpatialDomains::HexGeomUniquePtr hexGeom = CreateHex(v, segVec, faceVec);
+
+    Nektar::LibUtilities::PointsType quadPointsTypeDir1 =
+        Nektar::LibUtilities::eGaussLobattoLegendre;
+    Nektar::LibUtilities::BasisType basisTypeDir1 =
+        Nektar::LibUtilities::eModified_A;
+    unsigned int numQuadPoints = 8;
+    unsigned int numModes      = 4;
+
+    const Nektar::LibUtilities::PointsKey quadPointsKeyDir1(numQuadPoints,
+                                                            quadPointsTypeDir1);
+    const Nektar::LibUtilities::BasisKey basisKeyDir1(basisTypeDir1, numModes,
+                                                      quadPointsKeyDir1);
+
+    Nektar::LocalRegions::HexExpSharedPtr Exp =
+        MemoryManager<Nektar::LocalRegions::HexExp>::AllocateSharedPtr(
+            basisKeyDir1, basisKeyDir1, basisKeyDir1, hexGeom.get());
+
+    Nektar::StdRegions::StdHexExpSharedPtr stdExp =
+        MemoryManager<Nektar::StdRegions::StdHexExp>::AllocateSharedPtr(
+            basisKeyDir1, basisKeyDir1, basisKeyDir1);
+
+    int nelmts = 10;
+
+    std::vector<StdRegions::StdExpansionSharedPtr> CollExp;
+    for (int i = 0; i < nelmts; ++i)
+    {
+        CollExp.push_back(Exp);
+    }
+
+    LibUtilities::SessionReaderSharedPtr dummySession;
+    Collections::CollectionOptimisation colOpt(dummySession, 2,
+                                               Collections::eMatrixFree);
+    Collections::OperatorImpMap impTypes = colOpt.GetOperatorImpMap(stdExp);
+    Collections::Collection c(CollExp, impTypes);
+    StdRegions::ConstFactorMap factors;
+    factors[StdRegions::eFactorLambda] = -1.5;
+
+    c.Initialise(Collections::eLinearAdvectionDiffusionReaction, factors);
+
+    // Add advection velocities via varcoeffs
+    int npoints = Exp->GetTotPoints() * nelmts;
+    StdRegions::VarCoeffMap varcoeffs;
+    StdRegions::VarCoeffType varcoefftypes[] = {StdRegions::eVarCoeffVelX,
+                                                StdRegions::eVarCoeffVelY,
+                                                StdRegions::eVarCoeffVelZ};
+    for (int i = 0; i < Exp->GetShapeDimension(); i++)
+    {
+        varcoeffs[varcoefftypes[i]] = Array<OneD, NekDouble>(npoints, 1.0);
+    }
+    c.UpdateVarcoeffs(Collections::eLinearAdvectionDiffusionReaction,
+                      varcoeffs);
+
+    const int nm = Exp->GetNcoeffs();
+    Array<OneD, NekDouble> coeffsIn(nelmts * nm);
+    Array<OneD, NekDouble> coeffsRef(nelmts * nm);
+    Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
+
+    for (int i = 0; i < coeffsIn.size(); ++i)
+    {
+        coeffsIn[i] = i + 1.0;
+    }
+
+    StdRegions::StdMatrixKey mkey(StdRegions::eLinearAdvectionDiffusionReaction,
+                                  Exp->DetShapeType(), *Exp, factors,
+                                  varcoeffs);
+
+    for (int i = 0; i < nelmts; ++i)
+    {
+        // Standard routines
+        Exp->GeneralMatrixOp(coeffsIn + i * nm, tmp = coeffsRef + i * nm, mkey);
+    }
+
+    c.ApplyOperator(Collections::eLinearAdvectionDiffusionReaction, coeffsIn,
+                    coeffs);
+
+    double epsilon = 1.0e-8;
+    for (int i = 0; i < coeffsRef.size(); ++i)
+    {
+        coeffsRef[i] = (std::abs(coeffsRef[i]) < 1e-14) ? 0.0 : coeffsRef[i];
+        coeffs[i]    = (std::abs(coeffs[i]) < 1e-14) ? 0.0 : coeffs[i];
+        BOOST_CHECK_CLOSE(coeffsRef[i], coeffs[i], epsilon);
+    }
+}
+
 #endif
 } // namespace Nektar::HexCollectionTests