+++ library/UnitTests/Collections/TestPyrCollection.cpp
@@ -3284,14 +3284,9 @@ BOOST_AUTO_TEST_CASE(TestPyrHelmholtz_IterPerExp_UniformP_ConstVarDiff)
     Array<OneD, NekDouble> coeffsRef(nelmts * nm);
     Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
 
-    for (int i = 0; i < nm; ++i)
+    for (int i = 0; i < coeffsIn.size(); ++i)
     {
-        coeffsIn[i] = 1.0;
-    }
-
-    for (int i = 1; i < nelmts; ++i)
-    {
-        Vmath::Vcopy(nm, coeffsIn, 1, tmp = coeffsIn + i * nm, 1);
+        coeffsIn[i] = i + 1.0;
     }
 
     StdRegions::StdMatrixKey mkey(StdRegions::eHelmholtz, Exp->DetShapeType(),
@@ -3392,14 +3387,9 @@ BOOST_AUTO_TEST_CASE(TestPyrHelmholtz_MatrixFree_UniformP)
     Array<OneD, NekDouble> coeffsRef(nelmts * nm);
     Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
 
-    for (int i = 0; i < nm; ++i)
+    for (int i = 0; i < coeffsIn.size(); ++i)
     {
-        coeffsIn[i] = 1.0;
-    }
-
-    for (int i = 1; i < nelmts; ++i)
-    {
-        Vmath::Vcopy(nm, coeffsIn, 1, tmp = coeffsIn + i * nm, 1);
+        coeffsIn[i] = i + 1.0;
     }
 
     StdRegions::StdMatrixKey mkey(StdRegions::eHelmholtz, Exp->DetShapeType(),
@@ -3500,14 +3490,9 @@ BOOST_AUTO_TEST_CASE(TestPyrHelmholtz_MatrixFree_Deformed_overInt)
     Array<OneD, NekDouble> coeffsRef(nelmts * nm);
     Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
 
-    for (int i = 0; i < nm; ++i)
+    for (int i = 0; i < coeffsIn.size(); ++i)
     {
-        coeffsIn[i] = 1.0;
-    }
-
-    for (int i = 1; i < nelmts; ++i)
-    {
-        Vmath::Vcopy(nm, coeffsIn, 1, tmp = coeffsIn + i * nm, 1);
+        coeffsIn[i] = i + 1.0;
     }
 
     StdRegions::StdMatrixKey mkey(StdRegions::eHelmholtz, Exp->DetShapeType(),
@@ -3614,14 +3599,9 @@ BOOST_AUTO_TEST_CASE(TestPyrHelmholtz_MatrixFree_UniformP_ConstVarDiff)
     Array<OneD, NekDouble> coeffsRef(nelmts * nm);
     Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
 
-    for (int i = 0; i < nm; ++i)
+    for (int i = 0; i < coeffsIn.size(); ++i)
     {
-        coeffsIn[i] = 1.0;
-    }
-
-    for (int i = 1; i < nelmts; ++i)
-    {
-        Vmath::Vcopy(nm, coeffsIn, 1, tmp = coeffsIn + i * nm, 1);
+        coeffsIn[i] = i + 1.0;
     }
 
     StdRegions::StdMatrixKey mkey(StdRegions::eHelmholtz, Exp->DetShapeType(),
@@ -3837,6 +3817,363 @@ BOOST_AUTO_TEST_CASE(TestPyrPhysInterp1DScaled_MatrixFree_UniformP_MultiElmt)
         phys1[i]        = (fabs(phys1[i]) < 1e-14) ? 0.0 : phys1[i];
         exact           = (fabs(exact) < 1e-14) ? 0.0 : exact;
         BOOST_CHECK_CLOSE(phys1[i], exact, epsilon);
+    }
+}
+
+BOOST_AUTO_TEST_CASE(
+    TestPyrLinearAdvectionDiffusionReaction_IterPerExp_UniformP_ConstVarDiff)
+{
+    SpatialDomains::PointGeomUniquePtr v0(
+        new SpatialDomains::PointGeom(3u, 0u, -1.0, -1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v1(
+        new SpatialDomains::PointGeom(3u, 1u, 1.0, -1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v2(
+        new SpatialDomains::PointGeom(3u, 2u, 1.0, 1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v3(
+        new SpatialDomains::PointGeom(3u, 3u, -1.0, 1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v4(
+        new SpatialDomains::PointGeom(3u, 4u, -1.0, -1.0, 1.0));
+
+    std::array<SpatialDomains::SegGeomUniquePtr, 8> segVec;
+    std::array<SpatialDomains::QuadGeomUniquePtr, 1> quadVec;
+    std::array<SpatialDomains::TriGeomUniquePtr, 4> triVec;
+    std::array<SpatialDomains::PointGeom *, 5> v = {
+        v0.get(), v1.get(), v2.get(), v3.get(), v4.get()};
+    SpatialDomains::PyrGeomUniquePtr pyrGeom =
+        CreatePyr(v, segVec, triVec, quadVec);
+
+    unsigned int numQuadPoints = 5;
+    unsigned int numModes      = 4;
+
+    Nektar::LibUtilities::PointsType PointsTypeDir1 =
+        Nektar::LibUtilities::eGaussLobattoLegendre;
+    const Nektar::LibUtilities::PointsKey PointsKeyDir1(numQuadPoints,
+                                                        PointsTypeDir1);
+    Nektar::LibUtilities::BasisType basisTypeDir1 =
+        Nektar::LibUtilities::eModified_A;
+    const Nektar::LibUtilities::BasisKey basisKeyDir1(basisTypeDir1, numModes,
+                                                      PointsKeyDir1);
+
+    Nektar::LibUtilities::PointsType PointsTypeDir2 =
+        Nektar::LibUtilities::eGaussLobattoLegendre;
+    const Nektar::LibUtilities::PointsKey PointsKeyDir2(numQuadPoints,
+                                                        PointsTypeDir2);
+    Nektar::LibUtilities::BasisType basisTypeDir2 =
+        Nektar::LibUtilities::eModified_A;
+    const Nektar::LibUtilities::BasisKey basisKeyDir2(basisTypeDir2, numModes,
+                                                      PointsKeyDir2);
+
+    Nektar::LibUtilities::PointsType PointsTypeDir3 =
+        Nektar::LibUtilities::eGaussRadauMAlpha2Beta0;
+    const Nektar::LibUtilities::PointsKey PointsKeyDir3(numQuadPoints - 1,
+                                                        PointsTypeDir3);
+    Nektar::LibUtilities::BasisType basisTypeDir3 =
+        Nektar::LibUtilities::eModifiedPyr_C;
+    const Nektar::LibUtilities::BasisKey basisKeyDir3(basisTypeDir3, numModes,
+                                                      PointsKeyDir3);
+
+    Nektar::LocalRegions::PyrExpSharedPtr Exp =
+        MemoryManager<Nektar::LocalRegions::PyrExp>::AllocateSharedPtr(
+            basisKeyDir1, basisKeyDir2, basisKeyDir3, pyrGeom.get());
+
+    int nelmts = 10;
+
+    std::vector<StdRegions::StdExpansionSharedPtr> CollExp;
+    for (int i = 0; i < nelmts; ++i)
+    {
+        CollExp.push_back(Exp);
+    }
+
+    LibUtilities::SessionReaderSharedPtr dummySession;
+    Collections::CollectionOptimisation colOpt(dummySession, 2,
+                                               Collections::eIterPerExp);
+    Collections::OperatorImpMap impTypes = colOpt.GetOperatorImpMap(Exp);
+    Collections::Collection c(CollExp, impTypes);
+    StdRegions::ConstFactorMap factors;
+    factors[StdRegions::eFactorLambda] = 1.5;
+
+    c.Initialise(Collections::eLinearAdvectionDiffusionReaction, factors);
+
+    // Add advection velocities via varcoeffs
+    int npoints = Exp->GetTotPoints() * nelmts;
+    StdRegions::VarCoeffMap varcoeffs;
+    StdRegions::VarCoeffType varcoefftypes[] = {StdRegions::eVarCoeffVelX,
+                                                StdRegions::eVarCoeffVelY,
+                                                StdRegions::eVarCoeffVelZ};
+    for (int i = 0; i < Exp->GetShapeDimension(); i++)
+    {
+        varcoeffs[varcoefftypes[i]] = Array<OneD, NekDouble>(npoints, 1.0);
+    }
+    c.UpdateVarcoeffs(Collections::eLinearAdvectionDiffusionReaction,
+                      varcoeffs);
+
+    const int nm = Exp->GetNcoeffs();
+    Array<OneD, NekDouble> coeffsIn(nelmts * nm);
+    Array<OneD, NekDouble> coeffsRef(nelmts * nm);
+    Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
+
+    for (int i = 0; i < coeffsIn.size(); ++i)
+    {
+        coeffsIn[i] = i + 1.0;
+    }
+
+    StdRegions::StdMatrixKey mkey(StdRegions::eLinearAdvectionDiffusionReaction,
+                                  Exp->DetShapeType(), *Exp, factors,
+                                  varcoeffs);
+
+    for (int i = 0; i < nelmts; ++i)
+    {
+        // Standard routines
+        Exp->GeneralMatrixOp(coeffsIn + i * nm, tmp = coeffsRef + i * nm, mkey);
+    }
+
+    c.ApplyOperator(Collections::eLinearAdvectionDiffusionReaction, coeffsIn,
+                    coeffs);
+
+    double epsilon = 1.0e-8;
+    for (int i = 0; i < coeffsRef.size(); ++i)
+    {
+        coeffsRef[i] = (std::abs(coeffsRef[i]) < 1e-14) ? 0.0 : coeffsRef[i];
+        coeffs[i]    = (std::abs(coeffs[i]) < 1e-14) ? 0.0 : coeffs[i];
+        BOOST_CHECK_CLOSE(coeffsRef[i], coeffs[i], epsilon);
+    }
+}
+
+BOOST_AUTO_TEST_CASE(
+    TestPyrLinearAdvectionDiffusionReaction_MatrixFree_UniformP)
+{
+    SpatialDomains::PointGeomUniquePtr v0(
+        new SpatialDomains::PointGeom(3u, 0u, -1.0, -1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v1(
+        new SpatialDomains::PointGeom(3u, 1u, 1.0, -1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v2(
+        new SpatialDomains::PointGeom(3u, 2u, 1.0, 1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v3(
+        new SpatialDomains::PointGeom(3u, 3u, -1.0, 1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v4(
+        new SpatialDomains::PointGeom(3u, 4u, -1.0, -1.0, 1.0));
+
+    std::array<SpatialDomains::SegGeomUniquePtr, 8> segVec;
+    std::array<SpatialDomains::QuadGeomUniquePtr, 1> quadVec;
+    std::array<SpatialDomains::TriGeomUniquePtr, 4> triVec;
+    std::array<SpatialDomains::PointGeom *, 5> v = {
+        v0.get(), v1.get(), v2.get(), v3.get(), v4.get()};
+    SpatialDomains::PyrGeomUniquePtr pyrGeom =
+        CreatePyr(v, segVec, triVec, quadVec);
+
+    unsigned int numQuadPoints = 5;
+    unsigned int numModes      = 4;
+
+    Nektar::LibUtilities::PointsType PointsTypeDir1 =
+        Nektar::LibUtilities::eGaussLobattoLegendre;
+    const Nektar::LibUtilities::PointsKey PointsKeyDir1(numQuadPoints,
+                                                        PointsTypeDir1);
+    Nektar::LibUtilities::BasisType basisTypeDir1 =
+        Nektar::LibUtilities::eModified_A;
+    const Nektar::LibUtilities::BasisKey basisKeyDir1(basisTypeDir1, numModes,
+                                                      PointsKeyDir1);
+
+    Nektar::LibUtilities::PointsType PointsTypeDir2 =
+        Nektar::LibUtilities::eGaussLobattoLegendre;
+    const Nektar::LibUtilities::PointsKey PointsKeyDir2(numQuadPoints,
+                                                        PointsTypeDir2);
+    Nektar::LibUtilities::BasisType basisTypeDir2 =
+        Nektar::LibUtilities::eModified_A;
+    const Nektar::LibUtilities::BasisKey basisKeyDir2(basisTypeDir2, numModes,
+                                                      PointsKeyDir2);
+
+    Nektar::LibUtilities::PointsType PointsTypeDir3 =
+        Nektar::LibUtilities::eGaussRadauMAlpha2Beta0;
+    const Nektar::LibUtilities::PointsKey PointsKeyDir3(numQuadPoints - 1,
+                                                        PointsTypeDir3);
+    Nektar::LibUtilities::BasisType basisTypeDir3 =
+        Nektar::LibUtilities::eModifiedPyr_C;
+    const Nektar::LibUtilities::BasisKey basisKeyDir3(basisTypeDir3, numModes,
+                                                      PointsKeyDir3);
+
+    Nektar::LocalRegions::PyrExpSharedPtr Exp =
+        MemoryManager<Nektar::LocalRegions::PyrExp>::AllocateSharedPtr(
+            basisKeyDir1, basisKeyDir2, basisKeyDir3, pyrGeom.get());
+
+    int nelmts = 10;
+
+    std::vector<StdRegions::StdExpansionSharedPtr> CollExp;
+    for (int i = 0; i < nelmts; ++i)
+    {
+        CollExp.push_back(Exp);
+    }
+
+    LibUtilities::SessionReaderSharedPtr dummySession;
+    Collections::CollectionOptimisation colOpt(dummySession, 2,
+                                               Collections::eMatrixFree);
+    Collections::OperatorImpMap impTypes = colOpt.GetOperatorImpMap(Exp);
+    Collections::Collection c(CollExp, impTypes);
+    StdRegions::ConstFactorMap factors;
+    factors[StdRegions::eFactorLambda] = 1.5;
+
+    c.Initialise(Collections::eLinearAdvectionDiffusionReaction, factors);
+
+    // Add advection velocities via varcoeffs
+    int npoints = Exp->GetTotPoints() * nelmts;
+    StdRegions::VarCoeffMap varcoeffs;
+    StdRegions::VarCoeffType varcoefftypes[] = {StdRegions::eVarCoeffVelX,
+                                                StdRegions::eVarCoeffVelY,
+                                                StdRegions::eVarCoeffVelZ};
+    for (int i = 0; i < Exp->GetShapeDimension(); i++)
+    {
+        varcoeffs[varcoefftypes[i]] = Array<OneD, NekDouble>(npoints, 1.0);
+    }
+    c.UpdateVarcoeffs(Collections::eLinearAdvectionDiffusionReaction,
+                      varcoeffs);
+
+    const int nm = Exp->GetNcoeffs();
+    Array<OneD, NekDouble> coeffsIn(nelmts * nm);
+    Array<OneD, NekDouble> coeffsRef(nelmts * nm);
+    Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
+
+    for (int i = 0; i < coeffsIn.size(); ++i)
+    {
+        coeffsIn[i] = i + 1.0;
+    }
+
+    StdRegions::StdMatrixKey mkey(StdRegions::eLinearAdvectionDiffusionReaction,
+                                  Exp->DetShapeType(), *Exp, factors,
+                                  varcoeffs);
+
+    for (int i = 0; i < nelmts; ++i)
+    {
+        // Standard routines
+        Exp->GeneralMatrixOp(coeffsIn + i * nm, tmp = coeffsRef + i * nm, mkey);
+    }
+
+    c.ApplyOperator(Collections::eLinearAdvectionDiffusionReaction, coeffsIn,
+                    coeffs);
+
+    double epsilon = 1.0e-8;
+    for (int i = 0; i < coeffsRef.size(); ++i)
+    {
+        coeffsRef[i] = (std::abs(coeffsRef[i]) < 1e-14) ? 0.0 : coeffsRef[i];
+        coeffs[i]    = (std::abs(coeffs[i]) < 1e-14) ? 0.0 : coeffs[i];
+        BOOST_CHECK_CLOSE(coeffsRef[i], coeffs[i], epsilon);
+    }
+}
+
+BOOST_AUTO_TEST_CASE(
+    TestPyrLinearAdvectionDiffusionReaction_MatrixFree_Deformed_overInt)
+{
+    SpatialDomains::PointGeomUniquePtr v0(
+        new SpatialDomains::PointGeom(3u, 0u, -1.5, -1.5, -1.5));
+    SpatialDomains::PointGeomUniquePtr v1(
+        new SpatialDomains::PointGeom(3u, 1u, 1.0, -1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v2(
+        new SpatialDomains::PointGeom(3u, 2u, 1.0, 1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v3(
+        new SpatialDomains::PointGeom(3u, 3u, -1.0, 1.0, -1.0));
+    SpatialDomains::PointGeomUniquePtr v4(
+        new SpatialDomains::PointGeom(3u, 4u, -1.0, -1.0, 1.0));
+
+    std::array<SpatialDomains::SegGeomUniquePtr, 8> segVec;
+    std::array<SpatialDomains::QuadGeomUniquePtr, 1> quadVec;
+    std::array<SpatialDomains::TriGeomUniquePtr, 4> triVec;
+    std::array<SpatialDomains::PointGeom *, 5> v = {
+        v0.get(), v1.get(), v2.get(), v3.get(), v4.get()};
+    SpatialDomains::PyrGeomUniquePtr pyrGeom =
+        CreatePyr(v, segVec, triVec, quadVec);
+
+    unsigned int numQuadPoints = 8;
+    unsigned int numModes      = 4;
+
+    Nektar::LibUtilities::PointsType PointsTypeDir1 =
+        Nektar::LibUtilities::eGaussLobattoLegendre;
+    const Nektar::LibUtilities::PointsKey PointsKeyDir1(numQuadPoints,
+                                                        PointsTypeDir1);
+    Nektar::LibUtilities::BasisType basisTypeDir1 =
+        Nektar::LibUtilities::eModified_A;
+    const Nektar::LibUtilities::BasisKey basisKeyDir1(basisTypeDir1, numModes,
+                                                      PointsKeyDir1);
+
+    Nektar::LibUtilities::PointsType PointsTypeDir2 =
+        Nektar::LibUtilities::eGaussLobattoLegendre;
+    const Nektar::LibUtilities::PointsKey PointsKeyDir2(numQuadPoints,
+                                                        PointsTypeDir2);
+    Nektar::LibUtilities::BasisType basisTypeDir2 =
+        Nektar::LibUtilities::eModified_A;
+    const Nektar::LibUtilities::BasisKey basisKeyDir2(basisTypeDir2, numModes,
+                                                      PointsKeyDir2);
+
+    Nektar::LibUtilities::PointsType PointsTypeDir3 =
+        Nektar::LibUtilities::eGaussRadauMAlpha2Beta0;
+    const Nektar::LibUtilities::PointsKey PointsKeyDir3(numQuadPoints - 1,
+                                                        PointsTypeDir3);
+    Nektar::LibUtilities::BasisType basisTypeDir3 =
+        Nektar::LibUtilities::eModifiedPyr_C;
+    const Nektar::LibUtilities::BasisKey basisKeyDir3(basisTypeDir3, numModes,
+                                                      PointsKeyDir3);
+
+    Nektar::LocalRegions::PyrExpSharedPtr Exp =
+        MemoryManager<Nektar::LocalRegions::PyrExp>::AllocateSharedPtr(
+            basisKeyDir1, basisKeyDir2, basisKeyDir3, pyrGeom.get());
+
+    int nelmts = 10;
+
+    std::vector<StdRegions::StdExpansionSharedPtr> CollExp;
+    for (int i = 0; i < nelmts; ++i)
+    {
+        CollExp.push_back(Exp);
+    }
+
+    LibUtilities::SessionReaderSharedPtr dummySession;
+    Collections::CollectionOptimisation colOpt(dummySession, 2,
+                                               Collections::eMatrixFree);
+    Collections::OperatorImpMap impTypes = colOpt.GetOperatorImpMap(Exp);
+    Collections::Collection c(CollExp, impTypes);
+    StdRegions::ConstFactorMap factors;
+    factors[StdRegions::eFactorLambda] = 1.5;
+
+    c.Initialise(Collections::eLinearAdvectionDiffusionReaction, factors);
+
+    // Add advection velocities via varcoeffs
+    int npoints = Exp->GetTotPoints() * nelmts;
+    StdRegions::VarCoeffMap varcoeffs;
+    StdRegions::VarCoeffType varcoefftypes[] = {StdRegions::eVarCoeffVelX,
+                                                StdRegions::eVarCoeffVelY,
+                                                StdRegions::eVarCoeffVelZ};
+    for (int i = 0; i < Exp->GetShapeDimension(); i++)
+    {
+        varcoeffs[varcoefftypes[i]] = Array<OneD, NekDouble>(npoints, 1.0);
+    }
+    c.UpdateVarcoeffs(Collections::eLinearAdvectionDiffusionReaction,
+                      varcoeffs);
+
+    const int nm = Exp->GetNcoeffs();
+    Array<OneD, NekDouble> coeffsIn(nelmts * nm);
+    Array<OneD, NekDouble> coeffsRef(nelmts * nm);
+    Array<OneD, NekDouble> coeffs(nelmts * nm), tmp;
+
+    for (int i = 0; i < coeffsIn.size(); ++i)
+    {
+        coeffsIn[i] = i + 1.0;
+    }
+
+    StdRegions::StdMatrixKey mkey(StdRegions::eLinearAdvectionDiffusionReaction,
+                                  Exp->DetShapeType(), *Exp, factors,
+                                  varcoeffs);
+
+    for (int i = 0; i < nelmts; ++i)
+    {
+        // Standard routines
+        Exp->GeneralMatrixOp(coeffsIn + i * nm, tmp = coeffsRef + i * nm, mkey);
+    }
+
+    c.ApplyOperator(Collections::eLinearAdvectionDiffusionReaction, coeffsIn,
+                    coeffs);
+
+    double epsilon = 1.0e-8;
+    for (int i = 0; i < coeffsRef.size(); ++i)
+    {
+        coeffsRef[i] = (std::abs(coeffsRef[i]) < 1e-14) ? 0.0 : coeffsRef[i];
+        coeffs[i]    = (std::abs(coeffs[i]) < 1e-14) ? 0.0 : coeffs[i];
+        BOOST_CHECK_CLOSE(coeffsRef[i], coeffs[i], epsilon);
     }
 }